refactor notifications module
